requirements.txt
Flask_Cors==3.0.10
Flask==2.2.2
Werkzeug==2.2.2
Pillow==9.3.0
PyMuPDF==1.21.0
